+++ src/fetch-client-ui/components/Collection/AddTo/addTo.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect } from "react";
 import { useState } from "react";
 import { v4 as uuidv4 } from 'uuid';
-import { requestTypes, responseTypes } from "../../../utils/configuration";
-import { formatDate } from "../../../utils/helper";
-import vscode from "../Common/vscodeAPI";
-import { InitialSettings } from "../SideBar/redux/reducer";
-import { ICollections, IFolder, IHistory } from "../SideBar/redux/types";
-import { getMethodClassName } from "../SideBar/util";
-import "./style.css";
+import { requestTypes, responseTypes } from "../../../../utils/configuration";
+import { formatDate } from "../../../../utils/helper";
+import vscode from "../../Common/vscodeAPI";
+import { InitialSettings } from "../../SideBar/redux/reducer";
+import { ICollections, IFolder, IHistory } from "../../SideBar/redux/types";
+import { getMethodClassName } from "../../SideBar/util";
+import "../style.css";
 
 const AddToCollection = () => {